Improved example "To move the whole tree into a subdirectory..." to not fail when early commits are empty.
Previously this example would fail if the oldest commit(s) in any filtered branch is/are empty (no files) because the index would not change, and the mv would fail with:

	mv: cannot stat <redacted>/index.new': No such file or directory

This commonly occurs with histories created from git-svn-clone.  The updated example checks whether the index file has been created before attempting the mv.  The empty commit is retained.

See http://stackoverflow.com/questions/7798142/error-combining-git-repositories-into-subdirs for an example and explanation.

Signed-off-by: Brett Randall <javabrett@gmail.com>